docshell/test/test_bug509055.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=509055
-->
<head>
  <title>Test for Bug 509055</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=509055">Mozilla Bug 509055</a>
<p id="display"></p>
<div id="status"></div>
<div id="content">
</div>
<pre id="test">
<script type="application/javascript;version=1.7">

/** Test for Bug 509055 **/

SimpleTest.waitForExplicitFinish();

var gGen;

function shortWait() {
  setTimeout(function() { gGen.next(); }, 0, false);
}

function onChildLoad(e) {
  if(gGen)
    gGen.next();
}

function runTest() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

  var popup = window.open("file_bug509055.html", "popup 0",
                           "height=200,width=200,location=yes," +
                           "menubar=yes,status=yes,toolbar=yes,dependent=yes");
  popup.onload = onChildLoad;
  yield; // wait for load

  // Not sure why this wait is necessary, but without it, the change to
  // location.hash below doesn't create a SHEntry or enable the back button.
  shortWait();
  yield;

  // Both setting location.hash and calling history.back() happen
  // synchronously, so there's no need to yield here.
  popup.location.hash = "#1";
  popup.history.back();

  popup.document.title = "Changed";

  // Wait for listeners to be notified of the title change.
  shortWait();
  yield;

  var sh = popup.QueryInterface(Components.interfaces.nsIInterfaceRequestor)
                .getInterface(Components.interfaces.nsIWebNavigation)
                .sessionHistory;

  // Get the title of the inner popup's current SHEntry 
  var sheTitle = sh.getEntryAtIndex(sh.index, false).title;
  is(sheTitle, "Changed", "SHEntry's title should change when we change.");

  popup.close();

  SimpleTest.finish();
  yield;
}

window.addEventListener('load', function() {
  gGen = runTest();
  gGen.next();
}, false);

</script>

</body>
</html>
